v2plugin test env

+++ install/v2plugin/startcontiv.sh
@@ -12,6 +12,9 @@ mkdir -p $log_dir
 mkdir -p /var/run/openvswitch
 mkdir -p /etc/openvswitch
 
+touch /tmp/restart_netmaster
+touch /tmp/restart_netplugin
+
 echo "V2 Plugin logs" > $BOOTUP_LOGFILE
 
 if [ $iflist == "" ]; then
@@ -59,27 +62,30 @@ set +e
 
 echo "Starting Netplugin " >> $BOOTUP_LOGFILE
 while true ; do
+  if [ -f /tmp/restart_netplugin ]; then
     echo "/netplugin $dbg_flag -plugin-mode=$plugin_mode $vxlan_port_cfg -vlan-if=$iflist -cluster-store=$cluster_store $ctrl_ip_cfg $vtep_ip_cfg" >> $BOOTUP_LOGFILE
     /netplugin $dbg_flag -plugin-mode=$plugin_mode $vxlan_port_cfg -vlan-if=$iflist -cluster-store=$cluster_store $ctrl_ip_cfg $vtep_ip_cfg &> $log_dir/netplugin.log
     echo "CRITICAL : Net Plugin has exited, Respawn in 5" >> $BOOTUP_LOGFILE
     mv $log_dir/netplugin.log $log_dir/netplugin.log.lastrun
     sleep 5
     echo "Restarting Netplugin " >> $BOOTUP_LOGFILE
+  fi
 done &
 
 if [ $plugin_role == "master" ]; then
     echo "Starting Netmaster " >> $BOOTUP_LOGFILE
     while  true ; do
+      if [ -f /tmp/restart_netmaster ]; then
         echo "/netmaster $dbg_flag -plugin-name=$plugin_name -cluster-mode=$plugin_mode -cluster-store=$cluster_store $listen_url_cfg $control_url_cfg" >> $BOOTUP_LOGFILE
         /netmaster $dbg_flag -plugin-name=$plugin_name -cluster-mode=$plugin_mode -cluster-store=$cluster_store $listen_url_cfg $control_url_cfg &> $log_dir/netmaster.log
         echo "CRITICAL : Net Master has exited, Respawn in 5s" >> $BOOTUP_LOGFILE
 	mv $log_dir/netmaster.log $log_dir/netmaster.log.lastrun
         sleep 5
         echo "Restarting Netmaster " >> $BOOTUP_LOGFILE
+      fi
     done &
 else
     echo "Not starting netmaster as plugin role is" $plugin_role >> $BOOTUP_LOGFILE
 fi
 
 while true; do sleep 1; done
-
